+++ lib/godmin.rb
@@ -2,15 +2,15 @@ require "bootstrap-sass"
 require "bootstrap_form"
 require "momentjs-rails"
 require "selectize-rails"
-require "godmin/application"
+require "godmin/application_controller"
 require "godmin/authentication"
 require "godmin/authorization"
-require "godmin/model"
 require "godmin/engine"
 require "godmin/paginator"
 require "godmin/rails"
 require "godmin/resolver"
-require "godmin/resource"
+require "godmin/resources/resource_controller"
+require "godmin/resources/resource_service"
 require "godmin/version"
 
 module Godmin
