--- lib/godmin/model/pagination.rb
@@ -1,20 +1,0 @@
-module Godmin
-  module Model
-    module Pagination
-      extend ActiveSupport::Concern
-
-      def apply_pagination(page_param, resources)
-        @paginator = Paginator.new(resources, per_page: per_page, current_page: page_param)
-        @paginator.paginate
-      end
-
-      def paginator
-        @paginator
-      end
-
-      def per_page
-        25
-      end
-    end
-  end
-end
